Update & rename dictionary dispatch function to run in message-mode
Previously, the function jr/gnus-select-dispatch-dictionary-on-group
was called as part of the gnus-select-group-hook in an effort to
select the correct local dictionary for a group.

The issue here was that as soon as one switched to a message buffer to
reply or follow up an article, the local dictionary would default back
to the global dictionary.

With this change, we update the name of the function to
jr/message-dispatch-ispell-dictionary-on-group and have it hooked to
run as part of message-mode.